Move input methods to InputManager and finish InputManager
