+++ aplication/Main.java
@@ -1,10 +1,7 @@
 package aplication;
 
-import cards.Card;
-import interactions.View;
-
 public class Main{
     public static void main(String[] args){
-        
+
     }
 }
