+++ interactions/InputManager.java
@@ -1,20 +1,54 @@
 package interactions;
 
-public class InputManager{
+import java.util.Scanner;
+
+public class InputManager {
     private View view;
-    
+
     public InputManager() {
         view = new View();
     }
 
+    public String getStringInput(String message) {
+        System.out.println(" " + message);
+        System.out.print(" ");
+        Scanner scannerFromUser = new Scanner(System.in);
+        String input = scannerFromUser.nextLine();
+        
+        return input;
+    }
+
+    public int getIntInput(String message) {
+        System.out.println("\n " + message);
+        System.out.print(" ");
+        Scanner scannerFromUser = new Scanner(System.in);
+        int input = scannerFromUser.nextInt();
+        
+        return input;
+    }
+
     public String askForName(String intro) {
         String name = "";
-
-        view.getStringInput(intro + "! What's your name?");
+        name = getStringInput(intro + ", what's your name?");
         while (name.length() < 1) {
-            view.getStringInput(intro + "! What's your name?");
+            name = getStringInput("Wrong input! What's your name?");
         }
 
         return name;
+    }
+
+    public int askForStatToCompare() {
+        
+        String[] listOfStats = new String[] {"Deaths", "Incubation period", "Infectivity", 
+                                             "Painfullness", "Panic level"};
+        view.print(listOfStats, "Available stats:");
+        int statNumber = getIntInput("What do you want to compare? Type the number.");
+        
+        while (statNumber < 1 || statNumber > listOfStats.length) {
+            statNumber = getIntInput(String.format("Wrong input! type the number between 1 and %d", 
+                                                    listOfStats.length));
+        }
+
+        return statNumber;
     }
 }
